Fix type in equality assertion
The following error was reported in one of the Integration CI test run:
```
    redpanda_controller_test.go:563:
        	Error Trace:	/work/operator/internal/controller/redpanda/redpanda_controller_test.go:563
        	Error:      	Not equal:
        	            	expected: int(1)
        	            	actual  : int32(1)
        	Test:       	TestIntegrationRedpandaController/TestConnectorsIntegration
```

Reference
https://buildkite.com/redpanda/redpanda-operator/builds/4568#01955b8a-f282-478e-a0bf-6f45930eaa57/347-802

+++ operator/internal/controller/redpanda/redpanda_controller_test.go
@@ -560,7 +560,7 @@ func (s *RedpandaControllerSuite) TestConnectorsIntegration() {
 	var deployments appsv1.DeploymentList
 	s.NoError(s.client.List(s.ctx, &deployments))
 	s.Len(deployments.Items, 1)
-	s.Equal(1, deployments.Items[0].Status.ReadyReplicas)
+	s.Equal(int32(1), deployments.Items[0].Status.ReadyReplicas)
 
 	s.deleteAndWait(rp)
 }
